GridFS refactoring
File organization
Removed warnings
Handled IO errors
